requirements.txt
# flask
# gunicorn
# langchain
# langchain-google-genai
# langchain-pinecone
# pinecone-client
# beautifulsoup4
# requests
# python-dotenv
# sentence-transformers
# langchain-community
flask
gunicorn
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
beautifulsoup4
requests
python-dotenv
langchain-community